Add files via upload
validUser is a php function that given the parameters ($loginType, $login) returns true if the login given is valid for the given loginType (Admin or User) and false otherwise.

validUserTest is a program that displays the result.  I used it to test the logic, since I was having trouble integrating validUser with the existing login.php file to test.  validUserTest can be deleted once validUser.php and login.php are working together correctly.